借阅请求功能实现，数据库更新，代码重构，自动化工具位置预留。
为实现借阅功能，更新了数据库。
重构了views文件中的代码。
准备加入自动化工具。
