+++ NoAppFolder/状态码说明.md
@@ -72,6 +72,10 @@
 
 服务器不支持实现请求所需要的功能。
 
+**5598** 
+
+表示由于平台某些规则的限制，请求失败。 
+
 **5599** 
 
 表示请求失败 服务器异常
